requirements.txt
streamlit
openai
pandas
folium
streamlit-folium
fpdf2